+++ tailwind.config.js
@@ -7,12 +7,25 @@ module.exports = {
   ],
   theme: {
     extend: {
+      colors: {
+        'link-text': '#9ca3af',
+        'link-hover': '#0ea5e9',
+      },
       backgroundImage: {
         "gradient-radial": "radial-gradient(var(--tw-gradient-stops))",
         "gradient-conic":
-          "conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))",
+        "conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))",
+      },
+      fontFamily: {
+        arial: ['Arial', 'sans-serif'], 
       },
     },
   },
   plugins: [],
 };
+
+
+
+
+
+
